don't string-match the type name in isFixedSizeObject
This is a mistake I pushed for, before realizing that the
name of the type will be different outside of the module.
This is therefore not the right approach. Instead, I opted
to do the same thing as with the other method: check if the
type has a custom function to implement this feature.