don't inline array deserialization
While this is a nice optimization in the case of a small array, this
will cause the comptime evaluation of very large arrays to fail or
require that the number of evaluated branches are set to a higher
limit. Let's not do that, it's not worth it and it has implicaitons
on the instruction cache.